feat(layers)!: add native frame layers
Replace frame-only documents with strict v2 layered documents and expose deterministic editing across the core, CLI, Studio, and agent session boundary.

BREAKING CHANGE: v1 documents using clips[].frames are no longer accepted; documents must use the v2 layered format.

+++ src/core/Grid.cpp
@@ -66,9 +66,9 @@ QSet<QChar> Grid::slotsUsed() const
     return used;
 }
 
-int Grid::drawnCount() const
+qint64 Grid::drawnCount() const
 {
-    int drawn = 0;
+    qint64 drawn = 0;
     for (QChar cell : m_cells) {
         if (cell != Empty)
             ++drawn;
